+++ src/vue/Details/Password/DetailField.vue
@@ -2,7 +2,11 @@
     <div>
         {{label}}
         <web :href="getLink" v-if="['url','email','file'].indexOf(type) !== -1">{{ getLinkLabel }}</web>
-        <span @mouseenter="showValue=true" @mouseout="showValue=false" :class="getSecretClass" v-if="type === 'secret'" @click="copyValue">{{ getSecretValue }}</span>
+        <span @mouseenter="showValue=true"
+              @mouseout="showValue=false"
+              :class="getSecretClass"
+              v-if="type === 'secret'"
+              @click="copyValue">{{ getSecretValue }}</span>
         <span v-if="type === 'text'" @click="copyValue">{{ value }}</span>
     </div>
 </template>
@@ -18,7 +22,7 @@
             Web
         },
         props     : {
-            label : {
+            label: {
                 type: String
             },
             type : {
@@ -36,25 +40,29 @@
         },
         computed  : {
             getSecretValue() {
+                if (this.value === '') return '';
                 return this.showValue ? this.value:''.padStart(12, '⚫');
             },
             getSecretClass() {
                 return this.showValue ? 'secret visible':'secret';
             },
             getLinkLabel() {
-                if(this.type === 'file' && this.value.indexOf('/') !== -1) return this.value.substr(this.value.lastIndexOf('/') + 1);
+                if (this.type === 'file' && this.value.indexOf('/') !== -1) {
+                    return this.value.substr(this.value.lastIndexOf(
+                        '/') + 1);
+                }
                 return this.value;
             },
             getLink() {
-                if(this.type === 'email') return `mailto:${this.value}`;
-                if(this.type === 'file') return SettingsManager.get('server.baseUrl.webdav') + this.value.substr(1);
+                if (this.type === 'email') return `mailto:${this.value}`;
+                if (this.type === 'file') return SettingsManager.get('server.baseUrl.webdav') + this.value.substr(1);
                 return this.value;
             }
         },
         methods   : {
             copyValue() {
                 let message = 'Error copying {element} to clipboard';
-                if(Utility.copyToClipboard(this.value)) message = '{element} was copied to clipboard';
+                if (Utility.copyToClipboard(this.value)) message = '{element} was copied to clipboard';
                 Messages.notification([message, {element: this.label}]);
             }
         }